terraform/cloud-functions/metadata-generator/requirements.txt
google-cloud-aiplatform==1.106.0
google-cloud-firestore==2.21.0
google-cloud-pubsub==2.31.1
pytz==2025.2
python-dotenv==1.1.1
Jinja2==3.1.6